+++ demo/demo.js
@@ -70,6 +70,9 @@ if (useQuickStart) {
   core = GbElementsLogic.quickStart({
     customerId: customerId,
     productTransformer: productTransformer,
+    pluginOptions: {
+      gb_tracker: { area: 'DemoMaster' }
+    }
   });
 } else {
   // init core and plugins
@@ -77,12 +80,13 @@ if (useQuickStart) {
   const cachePlugin = new GbElementsLogic.CachePlugin();
   const cacheDriverPlugin = new GbElementsLogic.CacheDriverPlugin();
   const domEventsPlugin = new GbElementsLogic.DomEventsPlugin();
+  const gbTrackerPlugin = new GbElementsPlugins.GbTrackerPlugin({ customerId: 'apparel', area: 'DemoMaster' });
   const saytPlugin = new GbElementsLogic.SaytPlugin({ subdomain: customerId });
   const saytDriverPlugin = new GbElementsLogic.SaytDriverPlugin({ productTransformer: productTransformer });
   const searchPlugin = new GbElementsLogic.SearchPlugin({ customerId: customerId });
   const searchDriverPlugin = new GbElementsLogic.SearchDriverPlugin({ productTransformer: searchProductTransformer });
   // register all plugins with core
-  core.register([cachePlugin, cacheDriverPlugin, saytPlugin, saytDriverPlugin, domEventsPlugin, searchPlugin, searchDriverPlugin]);
+  core.register([cachePlugin, cacheDriverPlugin, saytPlugin, saytDriverPlugin, domEventsPlugin, searchPlugin, searchDriverPlugin, gbTrackerPlugin]);
 }
 
 [
@@ -92,6 +96,7 @@ if (useQuickStart) {
   'gbe::search_response',
   'gbe::sayt_products_error',
   'gbe::update_search_term',
+  'gbe::beacon_search',
 ].forEach(function(eventName) {
   window.addEventListener(eventName, function(event) {
     console.log(event.type, event.detail);