+++ packages/web-components/@groupby/elements-sayt/test/unit/sayt.test.ts
@@ -676,6 +676,7 @@ describe('Sayt Component', () => {
             area,
             collection,
           },
+          origin: 'sayt',
         },
         bubbles: true,
       });