+++ packages/web-components/@groupby/elements-sayt/CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased] [minor]
+### Added
+- ELE-249: Added an `origin` of `'sayt'` in the event payload requesting an update to the search term.
+
 ## [0.1.0] - 2019-11-29
 ### Added
 - SFX-191: Added the `sayt` component.